Feature/20200506 df tr wrk cg paper open (#8)
* Fixing createDir bug in AbstractRunner.

* Formatting & dbg fixes.

* Logic simplifications by CLion (double-check b/f deleting commented code).

* Running simulation using optzd values.

* Launching sim using optmzd case in standalone dir.

* Removing default value in inherited function.

* Added OPTMZD_DIR to paths.

* Initialized unset results_ and model_ vars.

* Dbg and formatting changes.

* Fix opt case dir for serial and mpi runners.

* Equal value check for continuous property.

* Changing timestep to ContinuousProperty double.

* Formatting & proposed fixes.

* Fix for constraint_handler in PSO optimizer.

* Refactoring.

* Adjusting TR parameters + refactoring: json_parameters -> json_params + add E error function to Printer::.

* Setting up debug version of TR-DFO.

* Fixing dbg output, some formatting.

* Abbreviated opt performance flags to ITERS, MIN, MAX.

* Renamed several TR-DFO params + changes to fit debug TR-DFO version (DFTR).

* Added prob_name parameter + changed to new TR-DFO param names.

* CMake update.

* Removing comments/obsolete code. Making sure all test in Optimization pass (OK).

+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/welspecs.h
@@ -59,7 +59,7 @@ class Welspecs : public ECLDriverPart
    * @param wells List of all wells.
    * @param timestep Timestep to generate data for.
    */
-  Welspecs(QList<Model::Wells::Well *> *wells, int timestep);
+  Welspecs(QList<Model::Wells::Well *> *wells, double timestep);
 
   QString GetPartString() const;
   string GetPartStdStr() const;
